refactored the delete route to use token

+++ controllers/blogs.js
@@ -14,9 +14,7 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const token = request.body.token
-
   const decodedToken = jwt.verify(token, process.env.SECRET)
-
 
   if(!token || !decodedToken.id){
     return response.status(401).json({ error: 'token missing or invalid' })
@@ -24,7 +22,6 @@ blogsRouter.post('/', async (request, response) => {
 
   const body = request.body;
   const user = await User.findById(decodedToken.id);
-
 
   const blog = new Blog({
     title: body.title,
@@ -56,9 +53,23 @@ blogsRouter.post('/', async (request, response) => {
   handleLikes()
   handleTitleAndUrl()
   
-})
+});
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.body.token;
+  const decodedToken = jwt.verify(request.body.token, process.env.SECRET)
+
+  if(!token || !decodedToken.id){
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id)
+
+  if( !(user.id.toString() === blog.user.toString()) ){
+    return response.status(401).json({ error: 'This blog can only be deleted by the user that entered it' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end()
 })
